bug fixes
git-svn-id: https://subversion.assembla.com/svn/smart-cs/trunk@5711 37136b1e-4cb1-4992-8394-573b48a9bdce

+++ source/extensions/i2/org.wcs.smart.i2/src/org/wcs/smart/i2/handlers/ConservationAreaCloner.java
@@ -133,10 +133,12 @@ public class ConservationAreaCloner implements IConservationAreaTemplateCloner{
 			clone.setIcon(ia.getIcon());
 			clone.setIdAttribute((IntelAttribute)engine.getNewConservationItem(ia.getIdAttribute()));
 			if (ia.getBirtTemplate() != null){
-				clone.setBirtTemplate(ia.getBirtTemplate());
 				Path source = IntelReportManager.INSTANCE.getEntityTemplate(ia);
 				Path target = IntelReportManager.INSTANCE.getEntityTemplate(clone);
-				FileUtils.copyFile(source.toFile(), target.toFile());
+				if (Files.exists(source)){
+					clone.setBirtTemplate(ia.getBirtTemplate());
+					FileUtils.copyFile(source.toFile(), target.toFile());
+				}
 			}
 			
 			engine.getSession().save(clone);
